+++ quantrl/__init__.py
@@ -1,2 +1,2 @@
 """Module to initialize QuantRL."""
-__version__ = "0.0.10"
+__version__ = "0.1.0"